automatic area height expansion
Added the ability to expand/shrink the typing area whilst the user is
typing. The code will add a minHeight that will be used, and will
adjust the height of the area as the user adds or removes text. If no
min-height is declared then the textAngular component will ignore this
functionality.

To use this functionality follow this example:
<div text-angular ng-model="htmlVariable" min-height=200></div>

Note: the height must be a number