Popup required 2 clicks to hide
After accessing a popup (IE image selecting), hiding the popup would require 2 clicks within the editor to close.  As a result this would cause havoc when clicking once outside the popup and then back on the image again, as the popup would not always come back or require multiple clicks on image and outside of image to comeback.  Solution, remove the animate remove class on hide popup, and just straight remove the class on hide since the animate removeClass only occurs on the next digest which only occurs on second click.