feat(taRegisterTool('indent'), taRegisterTool('outdent')): Enhanced toolbar to handle TAB and SHIFT-TAB KEYS.
The _keydown() function in main.js now detets the keyCode:9 (the TAB key) and keyCode:9 + shiftKey.
These two key events cause a specialKey property to be added to the key event:
  TAB key -> specialKey: 'TabKey'
  Shift TAB key -> specialKey: 'ShiftTabKey'
which is then passed on to _keypress().
taRegisterTool('indent', ...) is enhanced with a
 commandKeyCode: 'TabKey' -- so that the Tab key triggers an indent
taRegisterTool('outdent', ...) is enhanced with a
 commandKeyCode: 'ShiftTabKey' -- so that the Shift-Tab key triggers an outdent.
The textAngularManager::sendKeyCommand() detects the event.specialKey and then calls
 the registered tools.

Closes #734 and pull request #742